refactor: move configuration file to i-a main folder
