+++ src/sp-proxy/src/interface-adapters/data/FileReadProxyConfig.ts
@@ -1,6 +1,6 @@
 import { SpProxyConfig } from '@sp-proxy/entities/SpProxyConfig'
 import { IReadProxyConfigGateway } from '@sp-proxy/use-cases/ports/IReadProxyConfigGateway'
-import cfg from '@sp-proxy/interface-adapters/data/config/env'
+import cfg from '@sp-proxy/interface-adapters/config/env'
 import { readFileSync } from 'fs'
 import { SpProxyConfigProps } from '@sp-proxy/entities/protocols/SpProxyConfigProps'
 import { PersistenceError } from '@sp-proxy/interface-adapters/data/errors/PersistenceError'
